requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
pytest==8.3.4
requests==2.32.3 
httpx==0.28.1
mangum==0.19.0
boto3==1.36.12
watchtower==3.3.1
pydantic
python-dateutil
python-dotenv
starlette
pytz